+++ cpp/libjoynr/include/joynr/PublicationManager.h
@@ -84,7 +84,6 @@ public:
     PublicationManager(boost::asio::io_service& ioService,
                        IJoynrMessageSender* messageSender,
                        int maxThreads = 1);
-    PublicationManager(DelayedScheduler* scheduler, IJoynrMessageSender* messageSender);
     virtual ~PublicationManager();
     /**
      * @brief Adds the SubscriptionRequest and starts runnable to poll attributes.
@@ -228,7 +227,7 @@ private:
 
     std::mutex fileWriteLock;
     // Publications are scheduled to run on a thread pool
-    DelayedScheduler* delayedScheduler;
+    std::unique_ptr<DelayedScheduler> delayedScheduler;
 
     // Support for clean shutdowns
     std::mutex shutDownMutex;