+++ cpp/tests/unit-tests/PublicationManagerTest.cpp
@@ -116,7 +116,7 @@ TEST_F(PublicationManagerTest, add_requestCallerIsCalledCorrectlyByPublisherRunn
     InterfaceRegistrar::instance().registerRequestInterpreter<joynr::tests::testRequestInterpreter>(joynr::tests::testProvider::INTERFACE_NAME());
 
     MockPublicationSender mockPublicationSender;
-    MockTestRequestCaller* mockTestRequestCaller = new MockTestRequestCaller(Between(3, 5));
+    auto requestCaller = std::make_shared<MockTestRequestCaller>(Between(3, 5));
 
     // NOTE: it depends on the timing and especially on the CPU load of
     // the current machine how often the publication is exectuted. Hence,
@@ -125,7 +125,6 @@ TEST_F(PublicationManagerTest, add_requestCallerIsCalledCorrectlyByPublisherRunn
                 sendSubscriptionPublicationMock(_,_,_,_))
             .Times(Between(3, 5));
 
-    std::shared_ptr<MockTestRequestCaller> requestCaller(mockTestRequestCaller);
     PublicationManager publicationManager(singleThreadedIOService.getIOService(), messageSender);
 
     //SubscriptionRequest
@@ -153,7 +152,7 @@ TEST_F(PublicationManagerTest, add_requestCallerIsCalledCorrectlyByPublisherRunn
 
 TEST_F(PublicationManagerTest, stop_publications) {
     MockPublicationSender mockPublicationSender;
-    MockTestRequestCaller* mockTestRequestCaller = new MockTestRequestCaller(AtMost(2));
+    auto requestCaller = std::make_shared<MockTestRequestCaller>(AtMost(2));
 
     // Register the request interpreter that calls the request caller
     InterfaceRegistrar::instance().registerRequestInterpreter<tests::testRequestInterpreter>("tests/Test");
@@ -162,7 +161,6 @@ TEST_F(PublicationManagerTest, stop_publications) {
                 sendSubscriptionPublicationMock(_,_,_,_))
             .Times(AtMost(2));
 
-    std::shared_ptr<MockTestRequestCaller> requestCaller(mockTestRequestCaller);
     PublicationManager publicationManager(singleThreadedIOService.getIOService(), messageSender);
 
     //SubscriptionRequest
@@ -196,7 +194,7 @@ TEST_F(PublicationManagerTest, stop_publications) {
 
 TEST_F(PublicationManagerTest, remove_all_publications) {
     MockPublicationSender mockPublicationSender;
-    MockTestRequestCaller* mockTestRequestCaller = new MockTestRequestCaller(AtMost(2));
+    auto requestCaller = std::make_shared<MockTestRequestCaller>(AtMost(2));
 
     // Register the request interpreter that calls the request caller
     InterfaceRegistrar::instance().registerRequestInterpreter<tests::testRequestInterpreter>("tests/Test");
@@ -205,7 +203,6 @@ TEST_F(PublicationManagerTest, remove_all_publications) {
                 sendSubscriptionPublicationMock(_,_,_,_))
             .Times(AtMost(2));
 
-    std::shared_ptr<MockTestRequestCaller> requestCaller(mockTestRequestCaller);
     PublicationManager publicationManager(singleThreadedIOService.getIOService(), messageSender);
 
     //SubscriptionRequest
@@ -236,7 +233,7 @@ TEST_F(PublicationManagerTest, add_onChangeSubscription) {
     InterfaceRegistrar::instance().registerRequestInterpreter<tests::testRequestInterpreter>("tests/Test");
 
     MockPublicationSender mockPublicationSender;
-    MockTestRequestCaller* mockTestRequestCaller = new MockTestRequestCaller();
+    auto requestCaller = std::make_shared<MockTestRequestCaller>();
 
     // The attribute will change to this value
     joynr::types::Localisation::GpsLocation attributeValue;
@@ -262,7 +259,7 @@ TEST_F(PublicationManagerTest, add_onChangeSubscription) {
     std::string attributeName = "Location";
     SubscriptionAttributeListener* attributeListener;
     EXPECT_CALL(
-                *mockTestRequestCaller,
+                *requestCaller,
                 registerAttributeListener(attributeName,_)
     )
             .Times(1)
@@ -270,12 +267,11 @@ TEST_F(PublicationManagerTest, add_onChangeSubscription) {
 
     // Expect a call to remove the on change subscription
     EXPECT_CALL(
-                *mockTestRequestCaller,
+                *requestCaller,
                 unregisterAttributeListener(attributeName,_)
     )
             .Times(1);
 
-    std::shared_ptr<MockTestRequestCaller> requestCaller(mockTestRequestCaller);
     PublicationManager publicationManager(singleThreadedIOService.getIOService(), messageSender);
 
     //SubscriptionRequest
@@ -305,7 +301,7 @@ TEST_F(PublicationManagerTest, add_onChangeWithNoExpiryDate) {
     InterfaceRegistrar::instance().registerRequestInterpreter<tests::testRequestInterpreter>("tests/Test");
 
     MockPublicationSender mockPublicationSender;
-    MockTestRequestCaller* mockTestRequestCaller = new MockTestRequestCaller();
+    auto requestCaller = std::make_shared<MockTestRequestCaller>();
 
     // The attribute will change to this value
     joynr::types::Localisation::GpsLocation attributeValue;
@@ -331,13 +327,12 @@ TEST_F(PublicationManagerTest, add_onChangeWithNoExpiryDate) {
     std::string attributeName("Location");
     SubscriptionAttributeListener* attributeListener;
 
-    EXPECT_CALL(*mockTestRequestCaller,registerAttributeListener(attributeName, _))
+    EXPECT_CALL(*requestCaller,registerAttributeListener(attributeName, _))
             .Times(1)
             .WillRepeatedly(testing::SaveArg<1>(&attributeListener));
 
-    EXPECT_CALL(*mockTestRequestCaller,unregisterAttributeListener(attributeName, _)).Times(1);
+    EXPECT_CALL(*requestCaller,unregisterAttributeListener(attributeName, _)).Times(1);
 
-    std::shared_ptr<MockTestRequestCaller> requestCaller(mockTestRequestCaller);
     PublicationManager publicationManager(singleThreadedIOService.getIOService(), messageSender);
 
     //SubscriptionRequest
@@ -375,7 +370,7 @@ TEST_F(PublicationManagerTest, add_onChangeWithMinInterval) {
     InterfaceRegistrar::instance().registerRequestInterpreter<tests::testRequestInterpreter>("tests/Test");
 
     MockPublicationSender mockPublicationSender;
-    MockTestRequestCaller* mockTestRequestCaller = new MockTestRequestCaller();
+    auto requestCaller = std::make_shared<MockTestRequestCaller>();
 
     // The attribute will change to this value
     joynr::types::Localisation::GpsLocation attributeValue;
@@ -401,13 +396,12 @@ TEST_F(PublicationManagerTest, add_onChangeWithMinInterval) {
     std::string attributeName("Location");
     SubscriptionAttributeListener* attributeListener;
 
-    EXPECT_CALL(*mockTestRequestCaller,registerAttributeListener(attributeName, _))
+    EXPECT_CALL(*requestCaller,registerAttributeListener(attributeName, _))
             .Times(1)
             .WillRepeatedly(testing::SaveArg<1>(&attributeListener));
 
-    EXPECT_CALL(*mockTestRequestCaller,unregisterAttributeListener(attributeName, _)).Times(1);
+    EXPECT_CALL(*requestCaller,unregisterAttributeListener(attributeName, _)).Times(1);
 
-    std::shared_ptr<MockTestRequestCaller> requestCaller(mockTestRequestCaller);
     PublicationManager publicationManager(singleThreadedIOService.getIOService(), messageSender);
 
     //SubscriptionRequest
@@ -444,9 +438,9 @@ TEST_F(PublicationManagerTest, attribute_add_withExistingSubscriptionId) {
 
     // two request interpreters for the first and second add-API call
     MockPublicationSender mockPublicationSender;
-    MockTestRequestCaller* mockTestRequestCaller = new MockTestRequestCaller();
+    auto requestCaller = std::make_shared<MockTestRequestCaller>();
 
-    MockTestRequestCaller* mockTestRequestCaller2 = new MockTestRequestCaller();
+    auto requestCaller2 = std::make_shared<MockTestRequestCaller>();
     MockPublicationSender mockPublicationSender2;
 
     // The attribute will change to this value
@@ -483,20 +477,17 @@ TEST_F(PublicationManagerTest, attribute_add_withExistingSubscriptionId) {
     )
             .Times(2);
 
-    EXPECT_CALL(*mockTestRequestCaller,registerAttributeListener(attributeName, _))
+    EXPECT_CALL(*requestCaller,registerAttributeListener(attributeName, _))
             .Times(1)
             .WillRepeatedly(testing::SaveArg<1>(&attributeListener));
 
-    EXPECT_CALL(*mockTestRequestCaller,unregisterAttributeListener(attributeName, _)).Times(1);
+    EXPECT_CALL(*requestCaller,unregisterAttributeListener(attributeName, _)).Times(1);
 
-    EXPECT_CALL(*mockTestRequestCaller2,registerAttributeListener(attributeName, _))
+    EXPECT_CALL(*requestCaller2,registerAttributeListener(attributeName, _))
             .Times(1)
             .WillRepeatedly(testing::SaveArg<1>(&attributeListener));
 
-    EXPECT_CALL(*mockTestRequestCaller2,unregisterAttributeListener(attributeName, _)).Times(1);
-
-    std::shared_ptr<MockTestRequestCaller> requestCaller(mockTestRequestCaller);
-    std::shared_ptr<MockTestRequestCaller> requestCaller2(mockTestRequestCaller2);
+    EXPECT_CALL(*requestCaller2,unregisterAttributeListener(attributeName, _)).Times(1);
 
     PublicationManager publicationManager(singleThreadedIOService.getIOService(), messageSender);
 
@@ -566,7 +557,7 @@ TEST_F(PublicationManagerTest, attribute_add_withExistingSubscriptionId_testQos_
     InterfaceRegistrar::instance().registerRequestInterpreter<tests::testRequestInterpreter>("tests/Test");
 
     MockPublicationSender mockPublicationSender;
-    MockTestRequestCaller* mockTestRequestCaller = new MockTestRequestCaller();
+    auto requestCaller = std::make_shared<MockTestRequestCaller>();
 
     // The attribute will change to this value
     joynr::types::Localisation::GpsLocation attributeValue;
@@ -591,13 +582,11 @@ TEST_F(PublicationManagerTest, attribute_add_withExistingSubscriptionId_testQos_
     )
             .Times(3);
 
-    EXPECT_CALL(*mockTestRequestCaller,registerAttributeListener(attributeName, _))
+    EXPECT_CALL(*requestCaller,registerAttributeListener(attributeName, _))
             .Times(2)
             .WillRepeatedly(testing::SaveArg<1>(&attributeListener));
 
-    EXPECT_CALL(*mockTestRequestCaller,unregisterAttributeListener(attributeName, _)).Times(2);
-
-    std::shared_ptr<MockTestRequestCaller> requestCaller(mockTestRequestCaller);
+    EXPECT_CALL(*requestCaller,unregisterAttributeListener(attributeName, _)).Times(2);
 
     PublicationManager publicationManager(singleThreadedIOService.getIOService(), messageSender);
 
@@ -655,7 +644,7 @@ TEST_F(PublicationManagerTest, attribtue_add_withExistingSubscriptionId_testQos_
     InterfaceRegistrar::instance().registerRequestInterpreter<tests::testRequestInterpreter>("tests/Test");
 
     MockPublicationSender mockPublicationSender;
-    MockTestRequestCaller* mockTestRequestCaller = new MockTestRequestCaller();
+    auto requestCaller = std::make_shared<MockTestRequestCaller>();
 
     // The attribute will change to this value
     joynr::types::Localisation::GpsLocation attributeValue;
@@ -680,13 +669,11 @@ TEST_F(PublicationManagerTest, attribtue_add_withExistingSubscriptionId_testQos_
     )
             .Times(3);
 
-    EXPECT_CALL(*mockTestRequestCaller,registerAttributeListener(attributeName, _))
+    EXPECT_CALL(*requestCaller,registerAttributeListener(attributeName, _))
             .Times(2)
             .WillRepeatedly(testing::SaveArg<1>(&attributeListener));
 
-    EXPECT_CALL(*mockTestRequestCaller,unregisterAttributeListener(attributeName, _)).Times(2);
-
-    std::shared_ptr<MockTestRequestCaller> requestCaller(mockTestRequestCaller);
+    EXPECT_CALL(*requestCaller,unregisterAttributeListener(attributeName, _)).Times(2);
 
     PublicationManager publicationManager(singleThreadedIOService.getIOService(), messageSender);
 
@@ -743,10 +730,10 @@ TEST_F(PublicationManagerTest, broadcast_add_withExistingSubscriptionId) {
     InterfaceRegistrar::instance().registerRequestInterpreter<tests::testRequestInterpreter>("tests/Test");
 
     MockPublicationSender mockPublicationSender;
-    MockTestRequestCaller* mockTestRequestCaller = new MockTestRequestCaller();
+    auto requestCaller = std::make_shared<MockTestRequestCaller>();
 
-    MockTestRequestCaller* mockTestRequestCaller2 = new MockTestRequestCaller();
     MockPublicationSender mockPublicationSender2;
+    auto requestCaller2 = std::make_shared<MockTestRequestCaller>();
 
     // The broacast will fire this value
     joynr::types::Localisation::GpsLocation broadcastValue;
@@ -785,20 +772,17 @@ TEST_F(PublicationManagerTest, broadcast_add_withExistingSubscriptionId) {
     )
             .Times(1);
 
-    EXPECT_CALL(*mockTestRequestCaller,registerBroadcastListener(broadcastName, _))
+    EXPECT_CALL(*requestCaller,registerBroadcastListener(broadcastName, _))
             .Times(1)
             .WillRepeatedly(testing::SaveArg<1>(&broadcastListener));
 
-    EXPECT_CALL(*mockTestRequestCaller,unregisterBroadcastListener(broadcastName, _)).Times(1);
+    EXPECT_CALL(*requestCaller,unregisterBroadcastListener(broadcastName, _)).Times(1);
 
-    EXPECT_CALL(*mockTestRequestCaller2,registerBroadcastListener(broadcastName, _))
+    EXPECT_CALL(*requestCaller2,registerBroadcastListener(broadcastName, _))
             .Times(1)
             .WillRepeatedly(testing::SaveArg<1>(&broadcastListener));
 
-    EXPECT_CALL(*mockTestRequestCaller2,unregisterBroadcastListener(broadcastName, _)).Times(1);
-
-    std::shared_ptr<MockTestRequestCaller> requestCaller(mockTestRequestCaller);
-    std::shared_ptr<MockTestRequestCaller> requestCaller2(mockTestRequestCaller2);
+    EXPECT_CALL(*requestCaller2,unregisterBroadcastListener(broadcastName, _)).Times(1);
 
     PublicationManager publicationManager(singleThreadedIOService.getIOService(), messageSender);
 
@@ -861,7 +845,7 @@ TEST_F(PublicationManagerTest, broadcast_add_withExistingSubscriptionId_testQos_
     InterfaceRegistrar::instance().registerRequestInterpreter<tests::testRequestInterpreter>("tests/Test");
 
     MockPublicationSender mockPublicationSender;
-    MockTestRequestCaller* mockTestRequestCaller = new MockTestRequestCaller();
+    auto requestCaller = std::make_shared<MockTestRequestCaller>();
 
     // The broacast will fire this value
     joynr::types::Localisation::GpsLocation broadcastValue;
@@ -889,13 +873,11 @@ TEST_F(PublicationManagerTest, broadcast_add_withExistingSubscriptionId_testQos_
     )
             .Times(2);
 
-    EXPECT_CALL(*mockTestRequestCaller,registerBroadcastListener(broadcastName, _))
+    EXPECT_CALL(*requestCaller,registerBroadcastListener(broadcastName, _))
             .Times(2)
             .WillRepeatedly(testing::SaveArg<1>(&broadcastListener));
 
-    EXPECT_CALL(*mockTestRequestCaller,unregisterBroadcastListener(broadcastName, _)).Times(2);
-
-    std::shared_ptr<MockTestRequestCaller> requestCaller(mockTestRequestCaller);
+    EXPECT_CALL(*requestCaller,unregisterBroadcastListener(broadcastName, _)).Times(2);
 
     PublicationManager publicationManager(singleThreadedIOService.getIOService(), messageSender);
 
@@ -944,7 +926,7 @@ TEST_F(PublicationManagerTest, broadcast_add_withExistingSubscriptionId_testQos_
     InterfaceRegistrar::instance().registerRequestInterpreter<tests::testRequestInterpreter>("tests/Test");
 
     MockPublicationSender mockPublicationSender;
-    MockTestRequestCaller* mockTestRequestCaller = new MockTestRequestCaller();
+    auto requestCaller = std::make_shared<MockTestRequestCaller>();
 
     // The broacast will fire this value
     joynr::types::Localisation::GpsLocation broadcastValue;
@@ -972,13 +954,11 @@ TEST_F(PublicationManagerTest, broadcast_add_withExistingSubscriptionId_testQos_
     )
             .Times(1);
 
-    EXPECT_CALL(*mockTestRequestCaller,registerBroadcastListener(broadcastName, _))
+    EXPECT_CALL(*requestCaller,registerBroadcastListener(broadcastName, _))
             .Times(2)
             .WillRepeatedly(testing::SaveArg<1>(&broadcastListener));
 
-    EXPECT_CALL(*mockTestRequestCaller,unregisterBroadcastListener(broadcastName, _)).Times(2);
-
-    std::shared_ptr<MockTestRequestCaller> requestCaller(mockTestRequestCaller);
+    EXPECT_CALL(*requestCaller,unregisterBroadcastListener(broadcastName, _)).Times(2);
 
     PublicationManager publicationManager(singleThreadedIOService.getIOService(), messageSender);
 
@@ -1031,7 +1011,7 @@ TEST_F(PublicationManagerTest, remove_onChangeSubscription) {
     InterfaceRegistrar::instance().registerRequestInterpreter<tests::testRequestInterpreter>("tests/Test");
 
     MockPublicationSender mockPublicationSender;
-    MockTestRequestCaller* mockTestRequestCaller = new MockTestRequestCaller();
+    auto requestCaller = std::make_shared<MockTestRequestCaller>();
 
     // A publication should never be sent
     EXPECT_CALL(
@@ -1049,13 +1029,12 @@ TEST_F(PublicationManagerTest, remove_onChangeSubscription) {
     std::string attributeName("Location");
     SubscriptionAttributeListener* attributeListener;
 
-    EXPECT_CALL(*mockTestRequestCaller,registerAttributeListener(attributeName, _))
+    EXPECT_CALL(*requestCaller,registerAttributeListener(attributeName, _))
             .Times(1)
             .WillRepeatedly(testing::SaveArg<1>(&attributeListener));
 
-    EXPECT_CALL(*mockTestRequestCaller,unregisterAttributeListener(attributeName, _)).Times(1);
+    EXPECT_CALL(*requestCaller,unregisterAttributeListener(attributeName, _)).Times(1);
 
-    std::shared_ptr<MockTestRequestCaller> requestCaller(mockTestRequestCaller);
     PublicationManager publicationManager(singleThreadedIOService.getIOService(), messageSender);
 
     //SubscriptionRequest
@@ -1174,7 +1153,7 @@ TEST_F(PublicationManagerTest, forwardProviderRuntimeExceptionToPublicationSende
     InterfaceRegistrar::instance().registerRequestInterpreter<tests::testRequestInterpreter>("tests/Test");
 
     MockPublicationSender mockPublicationSender;
-    std::shared_ptr<MockTestRequestCaller> requestCaller = std::make_shared<MockTestRequestCaller>();
+    auto requestCaller = std::make_shared<MockTestRequestCaller>();
 
     // The value will be fired by the broadcast
     auto expected = std::make_shared<exceptions::ProviderRuntimeException>(requestCaller->providerRuntimeExceptionTestMsg);
@@ -1225,7 +1204,7 @@ TEST_F(PublicationManagerTest, forwardMethodInvocationExceptionToPublicationSend
     InterfaceRegistrar::instance().registerRequestInterpreter<tests::testRequestInterpreter>("tests/Test");
 
     MockPublicationSender mockPublicationSender;
-    std::shared_ptr<MockTestRequestCaller> requestCaller = std::make_shared<MockTestRequestCaller>();
+    auto requestCaller = std::make_shared<MockTestRequestCaller>();
 
     // The value will be fired by the broadcast
     auto expected = std::make_shared<exceptions::MethodInvocationException>("unknown method name for interface test: getNotExistingAttribute");