+++ java/messaging/messaging-common/src/test/java/io/joynr/messaging/util/MulticastWildcardRegexFactoryTest.java
@@ -54,6 +54,18 @@ public class MulticastWildcardRegexFactoryTest {
     }
 
     @Test
+    public void testOnlySingleLevelWildcard() {
+        Pattern pattern = subject.createIdPattern("+");
+        assertNotNull(pattern);
+        assertEquals("[^/]+", pattern.pattern());
+        assertTrue(pattern.matcher("onelevelhere").matches());
+        assertFalse(pattern.matcher("one/two").matches());
+        assertFalse(pattern.matcher("/one").matches());
+        assertFalse(pattern.matcher("one/").matches());
+        assertFalse(pattern.matcher("/one/two").matches());
+    }
+
+    @Test
     public void testSingleWildcardInMiddle() {
         Pattern pattern = subject.createIdPattern("one/+/three");
         assertNotNull(pattern);
@@ -88,6 +100,18 @@ public class MulticastWildcardRegexFactoryTest {
         assertTrue(pattern.matcher("one/two/three/four").matches());
         assertTrue(pattern.matcher("one/two").matches());
         assertFalse(pattern.matcher("one/twothree").matches());
+        assertFalse(pattern.matcher("").matches());
+    }
+
+    @Test
+    public void testOnlyMultiWildcard() {
+        Pattern pattern = subject.createIdPattern("*");
+        assertNotNull(pattern);
+        assertEquals(".+", pattern.pattern());
+        assertTrue(pattern.matcher("one").matches());
+        assertTrue(pattern.matcher("one/two").matches());
+        assertTrue(pattern.matcher("one/two/three").matches());
+        assertFalse(pattern.matcher("").matches());
     }
 
     @Test(expected = JoynrIllegalStateException.class)